pyproject.toml
[tool.poetry]
name = "fisherapp"
version = "0.1.0"
description = "Everything needed for the development of the Fishing App for Abenakis Fishers."
authors = ["Nassim Massaudi <n.nassim4100@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = "^0.22.0"
fastapi = "^0.99.0"
sqlalchemy = "^2.0.17"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
